+++ release-notes.md
@@ -3,6 +3,7 @@ Version 1.1.2 (unreleased)
 * Added Vec4i.toVec3d()
 * Added Vec3d.set(Vector3f)
 * Added Quatd.set(Quaternion)
+* Added a Vec3d.xzy() swizzle method
 
 Version 1.1.1 
 --------------